Fix tests

+++ evalml/tests/pipeline_tests/test_component_graph.py
@@ -1323,7 +1323,7 @@ def test_component_graph_types_merge_mock(mock_rf_fit):
         ["column_2", "column_1_a", "column_1_b", "column_1_c", "column_1_d", "column_3"]
     )
     assert isinstance(mock_rf_fit.call_args[0][0].ww.logical_types["column_3"], Integer)
-    assert isinstance(mock_rf_fit.call_args[0][0].ww.logical_types["column_2"], Integer)
+    assert isinstance(mock_rf_fit.call_args[0][0].ww.logical_types["column_2"], Double)
 
 
 def test_component_graph_preserves_ltypes_created_during_pipeline_evaluation():
